backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi==0.110.0",
    "httpx==0.27.0",
    "langchain==0.1.16",
    "langchain-google-genai==1.0.3",
    "langchain-groq==0.1.3",
    "langchain-openai==0.1.3",
    "pydantic==2.6.4",
    "python-dotenv==1.0.1",
    "python-multipart==0.0.9",
    "uvicorn[standard]==0.29.0",
]
